usb_c: Refactor Sink common code
Move non-sink specific code into common code
