+++ subsys/usb/usb_c/usbc_tc_common_internal.h
@@ -13,7 +13,6 @@
 
 #include "usbc_timer.h"
 #include "usbc_stack.h"
-#include "usbc_tc.h"
 
 /**
  * @brief Type-C States
@@ -23,16 +22,16 @@ enum tc_state_t {
 	TC_CC_OPEN_SUPER_STATE,
 	/** Super state that applies Rd to the CC lines */
 	TC_CC_RD_SUPER_STATE,
-	/** Disabled State */
-	TC_DISABLED_STATE,
-	/** Error Recovery State */
-	TC_ERROR_RECOVERY_STATE,
 	/** Unattached Sink State */
 	TC_UNATTACHED_SNK_STATE,
 	/** Attach Wait Sink State */
 	TC_ATTACH_WAIT_SNK_STATE,
 	/** Attached Sink State */
 	TC_ATTACHED_SNK_STATE,
+	/** Disabled State */
+	TC_DISABLED_STATE,
+	/** Error Recovery State */
+	TC_ERROR_RECOVERY_STATE,
 
 	/** Number of TC States */
 	TC_STATE_COUNT
@@ -105,5 +104,30 @@ enum tc_state_t tc_get_state(const struct device *dev);
  * @param enable set true to enable Power Deliver
  */
 void tc_pd_enable(const struct device *dev, const bool enable);
+
+/**
+ * @brief This function must only be called in the subsystem init function.
+ *
+ * @param dev Pointer to the device structure for the driver instance.
+ */
+void tc_subsys_init(const struct device *dev);
+
+/**
+ * @brief Run the TC Layer state machine. This is called from the subsystems
+ *        port stack thread.
+ *
+ * @param dev Pointer to the device structure for the driver instance.
+ * @param dpm_request Device Policy Manager request
+ */
+void tc_run(const struct device *dev, int32_t dpm_request);
+
+/**
+ * @brief Checks if the TC Layer is in an Attached state
+ *
+ * @param dev Pointer to the device structure for the driver instance.
+ *
+ * @retval true if TC Layer is in an Attached state, else false
+ */
+bool tc_is_in_attached_state(const struct device *dev);
 
 #endif /* ZEPHYR_SUBSYS_USBC_TC_COMMON_INTERNAL_H_ */
